Added setActuatorData(). Actuator data is now transmitted if the autocaptain is in HIL mode. Throttle commands up & period are now both uint16s. Logic has been removed from the output blocks in code_gen.mdl so moved into either the RC controller or autonomous controller code. RC sensor input for the controller block is now obtained after HIL/real world data selection to match the autonomous controller sensor inputs. The getters in commProtocol.c have all been modified to be more compact and utilize the proper data types.

+++ Code/basic_model/clib/commProtocol.h
@@ -71,7 +71,7 @@ typedef struct tActuatorData {
 	unsigned short r_period;
 	unsigned char b_enable;
 	unsigned char b_direction;
-	unsigned int t_identifier;
+	unsigned long t_identifier;
 	unsigned char data[6];
 	unsigned char size;
 	unsigned char trigger;
@@ -113,10 +113,10 @@ unsigned char calculateChecksum(char* sentence, unsigned char size);
 
 void getSensorData(float* data);
 
-void getActuatorData(float* data);
+void getActuatorData(unsigned long* data);
 
 void getStateData(float* data);
 
-void getCommandData(float* data);
+void getCommandData(unsigned long* data);
